Use the Pkg infrastructure to run tests
Now that we have our code in packages we can use the `Pkg`
infrastructure to run our tests.  This simplifies the code to run the
tests and brings it in line with the Julia community standard approach
to testing.
