+++ Project.toml
@@ -4,11 +4,13 @@ authors = ["Climate Modeling Alliance"]
 version = "0.1.0"
 
 [deps]
-Coverage = "a2441757-f6aa-5fb2-8edb-039e3f45d037"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 ParametersType = "046d00ba-03b1-11e9-3cca-b5213a01defd"
 PlanetParameters = "50a4fe1a-031f-11e9-2493-0505e0da639e"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Utilities = "d528e2a8-1396-11e9-0475-bb99f588b544"
+
+[extras]
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Pkg", "Test"]
