+++ src/PlanetParameters/Project.toml
@@ -3,3 +3,9 @@ uuid = "50a4fe1a-031f-11e9-2493-0505e0da639e"
 
 [deps]
 ParametersType = "046d00ba-03b1-11e9-3cca-b5213a01defd"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
